+++ branches/Lesh/lpja2/src/tactical/xml_enemy_armour_drops.cpp
@@ -107,22 +107,22 @@ armourDropEndElementHandle(void *userData, const char *name)
 		else if(strcmp(name, "uiIndex") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curArmourDrop.uiIndex   = (UINT32) atol(pData->szCharData);
+			pData->curArmourDrop.uiIndex   = (UINT32) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "ubArmourClass") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curArmourDrop.ubArmourClass   = (UINT8) atol(pData->szCharData);
+			pData->curArmourDrop.ubArmourClass   = (UINT8) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "ubEnemyDropRate") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curArmourDrop.ubEnemyDropRate  = (UINT8) atol(pData->szCharData);
+			pData->curArmourDrop.ubEnemyDropRate  = (UINT8) atoi(pData->szCharData);
 		}
 		else if (strcmp(name, "ubMilitiaDropRate") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curArmourDrop.ubMilitiaDropRate  = (UINT8) atol(pData->szCharData);
+			pData->curArmourDrop.ubMilitiaDropRate  = (UINT8) atoi(pData->szCharData);
 		}
 
 		pData->maxReadDepth--;
