+++ branches/Lesh/lpja2/src/tactical/xml_enemy_misc_drops.cpp
@@ -108,22 +108,22 @@ miscDropEndElementHandle(void *userData, const char *name)
 		else if(strcmp(name, "uiIndex") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curMiscDrop.uiIndex   = (UINT32) atol(pData->szCharData);
+			pData->curMiscDrop.uiIndex   = (UINT32) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "usItemClass") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curMiscDrop.usItemClass   = (UINT32) atol(pData->szCharData);
+			pData->curMiscDrop.usItemClass   = (UINT32) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "ubEnemyDropRate") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curMiscDrop.ubEnemyDropRate  = (UINT8) atol(pData->szCharData);
+			pData->curMiscDrop.ubEnemyDropRate  = (UINT8) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "ubMilitiaDropRate") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curMiscDrop.ubMilitiaDropRate  = (UINT8) atol(pData->szCharData);
+			pData->curMiscDrop.ubMilitiaDropRate  = (UINT8) atoi(pData->szCharData);
 		}
 
 		pData->maxReadDepth--;
