+++ branches/Lesh/lpja2/src/tactical/xml_explosive.cpp
@@ -112,52 +112,52 @@ explosiveEndElementHandle(void *userData, const char *name)
 		else if(strcmp(name, "uiIndex") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curExplosive.uiIndex   = (UINT32) atol(pData->szCharData);
+			pData->curExplosive.uiIndex   = (UINT32) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "ubType") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curExplosive.ubType  = (UINT8) atol(pData->szCharData);
+			pData->curExplosive.ubType  = (UINT8) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "ubDamage") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curExplosive.ubDamage = (UINT8) atol(pData->szCharData);
+			pData->curExplosive.ubDamage = (UINT8) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "ubStunDamage") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curExplosive.ubStunDamage   = (UINT8) atol(pData->szCharData);
+			pData->curExplosive.ubStunDamage   = (UINT8) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "ubRadius") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curExplosive.ubRadius   = (UINT8) atol(pData->szCharData);
+			pData->curExplosive.ubRadius   = (UINT8) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "ubVolume") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curExplosive.ubVolume   = (UINT8) atol(pData->szCharData);
+			pData->curExplosive.ubVolume   = (UINT8) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "ubVolatility") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curExplosive.ubVolatility   = (UINT8) atol(pData->szCharData);
+			pData->curExplosive.ubVolatility   = (UINT8) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "ubAnimationID") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curExplosive.ubAnimationID   = (UINT8) atol(pData->szCharData);
+			pData->curExplosive.ubAnimationID   = (UINT8) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "ubStartRadius") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curExplosive.ubStartRadius   = (UINT8) atol(pData->szCharData);
+			pData->curExplosive.ubStartRadius   = (UINT8) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "ubDuration") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curExplosive.ubDuration   = (UINT8) atol(pData->szCharData);
+			pData->curExplosive.ubDuration   = (UINT8) atoi(pData->szCharData);
 		}
 
 		pData->maxReadDepth--;
