+++ branches/Lesh/lpja2/src/tactical/xml_incompatible_attachments.cpp
@@ -107,12 +107,12 @@ incompatibleattachmentEndElementHandle(void *userData, const char *name)
 		else if(strcmp(name, "incompatibleattachmentIndex") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curIncompatibleAttachment[1] = (UINT16) atol(pData->szCharData);
+			pData->curIncompatibleAttachment[1] = (UINT16) atoi(pData->szCharData);
 		}
 		else if(strcmp(name, "itemIndex") == 0)
 		{
 			pData->curElement = ELEMENT;
-			pData->curIncompatibleAttachment[0] = (UINT16) atol(pData->szCharData);
+			pData->curIncompatibleAttachment[0] = (UINT16) atoi(pData->szCharData);
 		}
 
 		pData->maxReadDepth--;
